Removed unnecessary Services, integrated them into handlers
